test:load testing implement and doc of readme for using
